refactor(gitlab): move code
