+++ lib/modules/platform/gitlab/utils.ts
@@ -1,9 +1,23 @@
+import url from 'node:url';
 import is from '@sindresorhus/is';
+import { CONFIG_GIT_URL_UNAVAILABLE } from '../../../constants/error-messages';
+import { logger } from '../../../logger';
+import { getEnv } from '../../../util/env';
+import * as hostRules from '../../../util/host-rules';
+import type { HttpResponse } from '../../../util/http/types';
+import { parseUrl } from '../../../util/url';
 import { getPrBodyStruct } from '../pr-body';
-import type { GitLabMergeRequest, GitlabPr } from './types';
+import type { GitUrlOption } from '../types';
+import type { GitLabMergeRequest, GitlabPr, RepoResponse } from './types';
 
 export const DRAFT_PREFIX = 'Draft: ';
 export const DRAFT_PREFIX_DEPRECATED = 'WIP: ';
+
+export const defaults = {
+  hostType: 'gitlab',
+  endpoint: 'https://gitlab.com/api/v4/',
+  version: '0.0.0',
+};
 
 export function prInfo(mr: GitLabMergeRequest): GitlabPr {
   const pr: GitlabPr = {
@@ -39,4 +53,64 @@ export function prInfo(mr: GitLabMergeRequest): GitlabPr {
   }
 
   return pr;
+}
+
+export function getRepoUrl(
+  repository: string,
+  gitUrl: GitUrlOption | undefined,
+  res: HttpResponse<RepoResponse>,
+): string {
+  if (gitUrl === 'ssh') {
+    if (!res.body.ssh_url_to_repo) {
+      throw new Error(CONFIG_GIT_URL_UNAVAILABLE);
+    }
+    logger.debug(`Using ssh URL: ${res.body.ssh_url_to_repo}`);
+    return res.body.ssh_url_to_repo;
+  }
+
+  const opts = hostRules.find({
+    hostType: defaults.hostType,
+    url: defaults.endpoint,
+  });
+  const env = getEnv();
+
+  if (
+    gitUrl === 'endpoint' ||
+    is.nonEmptyString(env.GITLAB_IGNORE_REPO_URL) ||
+    res.body.http_url_to_repo === null
+  ) {
+    if (res.body.http_url_to_repo === null) {
+      logger.debug('no http_url_to_repo found. Falling back to old behavior.');
+    }
+    if (env.GITLAB_IGNORE_REPO_URL) {
+      logger.warn(
+        'GITLAB_IGNORE_REPO_URL environment variable is deprecated. Please use "gitUrl" option.',
+      );
+    }
+
+    // TODO: null check (#22198)
+    const { protocol, host, pathname } = parseUrl(defaults.endpoint)!;
+    const newPathname = pathname.slice(0, pathname.indexOf('/api'));
+    const uri = url.format({
+      protocol:
+        /* v8 ignore next: should never happen */
+        protocol.slice(0, -1) || 'https',
+      // TODO: types (#22198)
+      auth: `oauth2:${opts.token!}`,
+      host,
+      pathname: `${newPathname}/${repository}.git`,
+    });
+    logger.debug(`Using URL based on configured endpoint, url:${uri}`);
+    return uri;
+  }
+
+  logger.debug(`Using http URL: ${res.body.http_url_to_repo}`);
+  const repoUrl = parseUrl(res.body.http_url_to_repo);
+  // should never happen, but bad tests are causing that
+  if (!repoUrl) {
+    return '';
+  }
+  repoUrl.username = 'oauth2';
+  repoUrl.password = opts.token!;
+  return repoUrl.toString();
 }
